+++ src/recipe/recipe.module.ts
@@ -3,7 +3,7 @@ import { RecipeService } from './recipe.service';
 import { RecipeController } from './recipe.controller';
 import { PrismaService } from '../prisma/prisma.service';
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
-import { RedisCacheModule } from '../cache/redisCache.module';
+import { RedisCacheModule } from '../cache/redis-cache.module';
 import { RecipeCacheService } from './recipe.cache.service';
 
 @Module({
